refactor(docker): docker manager를 모듈로 분리
